+++ braintree/payment_method.py
@@ -1,4 +1,5 @@
 import braintree
+from braintree.address import Address
 from braintree.resource import Resource
 from braintree.configuration import Configuration
 
@@ -10,6 +11,10 @@ class PaymentMethod(Resource):
     @staticmethod
     def find(payment_method_token):
         return Configuration.gateway().payment_method.find(payment_method_token)
+
+    @staticmethod
+    def update(payment_method_token, params):
+        return Configuration.gateway().payment_method.update(payment_method_token, params)
 
     @staticmethod
     def delete(payment_method_token):
@@ -25,6 +30,45 @@ class PaymentMethod(Resource):
             "customer_id",
             "payment_method_nonce",
             "token",
-            {"options": ["make_default"]}
+            "billing_address_id",
+            {"billing_address": Address.create_signature()},
+            {"options": [
+                "make_default",
+                "verify_card",
+                "fail_on_duplicate_payment_method",
+                "verification_merchant_account_id",
+                ]
+            }
         ]
         return signature
+
+    @staticmethod
+    def update_signature():
+        signature = [
+            "billing_address_id",
+            "cardholder_name",
+            "cvv",
+            "device_session_id",
+            "expiration_date",
+            "expiration_month",
+            "expiration_year",
+            "number",
+            "token",
+            "venmo_sdk_payment_method_code",
+            "device_data",
+            "fraud_merchant_id",
+            "payment_method_nonce",
+            {"options": [
+                "make_default",
+                "verify_card",
+                "verification_merchant_account_id",
+                "venmo_sdk_session"
+                ]
+            },
+            {"billing_address" :
+                Address.update_signature() +
+                [{"options": ["update_existing"]}]
+            }
+        ]
+        return signature
+