+++ braintree/paypal_account.py
@@ -22,3 +22,8 @@ class PayPalAccount(Resource):
             {"options": ["make_default"]}
         ]
         return signature
+
+    def __init__(self, gateway, attributes):
+        Resource.__init__(self, gateway, attributes)
+        if "subscriptions" in attributes:
+            self.subscriptions = [braintree.subscription.Subscription(gateway, subscription) for subscription in self.subscriptions]